+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.1 - Browser compat & Error handling - 2020/06/21
+* Added error handling for extreme offsets
+* Added one-time message for non-chrome browsers, as I have confirmed that Firefox will always have silence. Use chrome for the best experience!
+
 ## 1.1.0 - New settings & timing tweaks - 2020/06/21
 * Added tweakable timeout offsets for startup and new audio
 * Changed default startup timeout to 2 seconds, as it was breaking on larger scenes